Medications Table Updates
The medications table was completed when atopic dermatitis (AD) was diagnoses and when AD was the only diagnosis.

The current analysis requires a lot of excel work to sort and present the data as desired. There's got to be an easier way to format the data so the 30 med variables can be assessed and all PATWT can be estimated for each drug. This seems tough for Stata. Perhaps much easier in R.