Comorbidities, Reliability Indices for Meds, ORs Trouble
Assessed reliability indices for medications.

Comorbidity analysis and assessment of reliability indices.

Attempted logistic regrission univariate, but not enough PSUs in each CSTRATA. Likely because restricting to children, but it was possible for previous. Maybe restrict to more recent time period...Perhaps the PSU problems are because of earlier time points.